requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.5.1
django-oauth-toolkit==3.0.1
Pillow==11.1.0
celery==5.5.3
redis==6.2.0
django-filter==25.1
psycopg2-binary==2.9.10
qrcode[pil]==8.2
python-decouple==3.8
coreapi
drf-yasg

